// src/app/core/models/funeralEventDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface FuneralEventDto {
  id: string;
  eventName: string;
  eventDate: string;
  location: string;
  description: string;
}